+++ src/pages/feed/ui/index.ts
@@ -1,8 +1,7 @@
 export { default as FollowTab } from './FollowTab'
-export { default as CdCarousel } from './CdCarousel'
 export { default as FeedCdList } from './FeedCdList'
 export { default as FeedProfile } from './FeedProfile'
 export { default as CdPlayerLayout } from './layout/CdPlayerLayout'
-export { default as CdViewerLayout } from './layout/CdViewerLayout'
+export { default as CarouselItem } from './CarouselItem'
 export { default as FollowLayout } from './layout/FollowLayout'
 export { default as FeedCdListLayout } from './layout/FeedCdListLayout'
